requirements.txt
pandas==2.2.1
numpy==1.26.4
jupyter==1.0.0
apache-airflow==2.8.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
python-dotenv==1.0.1
